% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-launch.R
\name{app-launch}
\alias{app-launch}
\alias{runExplorePRM}
\alias{appExplorePRM}
\alias{server_prm}
\alias{ui_prm}
\title{Run App}
\usage{
runExplorePRM(WORK_DIR = NULL, SAVE_PLOTDATA = FALSE, CLEAN = TRUE, ...)

appExplorePRM()

server_prm(input, output, session)

ui_prm()
}
\arguments{
\item{WORK_DIR}{working directory to launch the app from. Default is NULL which will use the current working directory.}

\item{SAVE_PLOTDATA}{Default is FALSE. Whether to keep plot input datasets generated during app session. Useful for plot development.}

\item{CLEAN}{Default is False. Whether to delete the session output directory. Useful for debuging}

\item{...}{additional args to pass to shiny::runApp}

\item{input}{server function arg}

\item{output}{server function arg}

\item{session}{server function arg}
}
\description{
Functions to run the package shiny app
}
\section{Functions}{
\itemize{
\item \code{runExplorePRM()}: run Recommendations Model Tuning App

\item \code{appExplorePRM()}: returns app object for subsequent execution

\item \code{server_prm()}: server function for app

\item \code{ui_prm()}: UI function for app

}}
