% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{app-output}
\alias{app-output}
\alias{.plot_theme}
\alias{.plot_title_style}
\alias{.parse_fails}
\alias{plot_diagnostic_0}
\alias{plot_diagnostic_1}
\alias{plot_diagnostic_2}
\alias{plot_diagnostic_3}
\alias{plot_diagnostic_4}
\alias{process_rec_ep}
\title{App Outputs}
\usage{
.plot_theme(.colors)

.plot_title_style(ptitle, psubtitle, .colors)

.parse_fails(rec)

plot_diagnostic_0(pdata0, .colors)

plot_diagnostic_1(pdata1, .colors)

plot_diagnostic_2(pdata2, .colors)

plot_diagnostic_3(pdata3, .colors)

plot_diagnostic_4(pdata4, .colors)

process_rec_ep(rec)
}
\arguments{
\item{.colors}{colors returned by get_app_colors()}

\item{ptitle}{plot title}

\item{psubtitle}{plot subtitle}

\item{rec}{internal}

\item{pdata0}{plot dataset}

\item{pdata1}{plot dataset}

\item{pdata2}{plot dataset}

\item{pdata3}{plot dataset}

\item{pdata4}{plot dataset}
}
\description{
Functions to generate and prepare outputs presented by the package app.
}
\section{Functions}{
\itemize{
\item \code{.plot_theme()}: provides the base theme for plots

\item \code{.plot_title_style()}: Style the plot title/subtitle

\item \code{.parse_fails()}: Parse failed skus to delineate between failures and uncertain recs

\item \code{plot_diagnostic_0()}: diagnostic plot

\item \code{plot_diagnostic_1()}: diagnostic plot

\item \code{plot_diagnostic_2()}: diagnostic plot

\item \code{plot_diagnostic_3()}: diagnostic plot

\item \code{plot_diagnostic_4()}: diagnostic plot

\item \code{process_rec_ep()}: build plot data with result of rdscore::restock_rec_ep

}}
