% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-plots.R
\name{app-plots}
\alias{app-plots}
\alias{.plot_theme}
\alias{.plot_title_style}
\alias{.plot_diagnostic_0}
\alias{.plot_diagnostic_1}
\alias{.plot_diagnostic_2}
\alias{.plot_diagnostic_3}
\alias{.plot_diagnostic_4}
\title{App Outputs}
\usage{
.plot_theme()

.plot_title_style(ptitle, psubtitle)

.plot_diagnostic_0(pdata0)

.plot_diagnostic_1(pdata1)

.plot_diagnostic_2(pdata2)

.plot_diagnostic_3(pdata3)

.plot_diagnostic_4(pdata4)
}
\arguments{
\item{ptitle}{plot title}

\item{psubtitle}{plot subtitle}

\item{pdata0}{plot dataset}

\item{pdata1}{plot dataset}

\item{pdata2}{plot dataset}

\item{pdata3}{plot dataset}

\item{pdata4}{plot dataset}
}
\description{
Functions to generate and prepare outputs presented by the package app.
}
\section{Functions}{
\itemize{
\item \code{.plot_theme()}: provides the base theme for plots

\item \code{.plot_title_style()}: Style the plot title/subtitle

\item \code{.plot_diagnostic_0()}: diagnostic plot

\item \code{.plot_diagnostic_1()}: diagnostic plot

\item \code{.plot_diagnostic_2()}: diagnostic plot

\item \code{.plot_diagnostic_3()}: diagnostic plot

\item \code{.plot_diagnostic_4()}: diagnostic plot

}}
