% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-ui.R
\name{app-ui}
\alias{app-ui}
\alias{.ui_bootstrap_theme}
\alias{.ui_selection_filters}
\alias{.ui_selection_stats}
\alias{.ui_model_plot_panel}
\alias{.ui_model_inputs_panel}
\alias{.slider_model_sales}
\alias{.slider_model_stock}
\alias{.slider_price_qrtls}
\alias{.slider_product_cls}
\alias{.slider_menu_period}
\alias{.toggle_model_ttest}
\title{Ui elements of MLTuning App}
\usage{
.ui_bootstrap_theme()

.ui_selection_filters()

.ui_selection_stats()

.ui_model_plot_panel()

.ui_model_inputs_panel()

.slider_model_sales()

.slider_model_stock()

.slider_price_qrtls()

.slider_product_cls()

.slider_menu_period()

.toggle_model_ttest()
}
\description{
Ui elements of MLTuning App
}
\section{Functions}{
\itemize{
\item \code{.ui_bootstrap_theme()}: bootstrap theme for app

\item \code{.ui_selection_filters()}: statistics panel

\item \code{.ui_selection_stats()}: statistics panel

\item \code{.ui_model_plot_panel()}: model diagnostic plot panel

\item \code{.ui_model_inputs_panel()}: model inputs panel

\item \code{.slider_model_sales()}: slider

\item \code{.slider_model_stock()}: slider

\item \code{.slider_price_qrtls()}: slider

\item \code{.slider_product_cls()}: slider

\item \code{.slider_menu_period()}: slider

\item \code{.toggle_model_ttest()}: ml toggle

}}
