% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{app-utils}
\alias{app-utils}
\alias{get_app_dir}
\alias{get_app_colors}
\alias{create_session_dir}
\alias{clean_session_dir}
\alias{save_plot_data}
\alias{create_dl_link}
\alias{generate_report}
\alias{save_plot_image}
\alias{build_scenario_data}
\alias{save_ml_context}
\alias{save_ml_results}
\alias{exec_ml_restock}
\alias{default_ml_params}
\alias{scale_ml_params}
\alias{unscale_ml_params}
\title{App Utils}
\usage{
get_app_dir()

get_app_colors()

create_session_dir()

clean_session_dir()

save_plot_data(results)

create_dl_link(dlfile = "report.html")

generate_report(org, store)

save_plot_image(p0, p1, p2, p3, p4)

build_scenario_data()

save_ml_context(oid, sid, sku_data, ml_args)

save_ml_results(results)

exec_ml_restock(oid, sid, sku_vec, ml_args)

default_ml_params()

scale_ml_params(ll)

unscale_ml_params(ll)
}
\arguments{
\item{results}{results from shiny app}

\item{dlfile}{download filename}

\item{org}{org name}

\item{store}{store name}

\item{p0}{plot}

\item{p1}{plot}

\item{p2}{plot}

\item{p3}{plot}

\item{p4}{plot}

\item{oid}{org id}

\item{sid}{store id}

\item{sku_data}{sku table}

\item{ml_args}{model args}

\item{sku_vec}{skus to get recs for}

\item{ll}{argument list to scale or unscale values}
}
\description{
Functions required to execute and facililate an application user session
}
\section{Functions}{
\itemize{
\item \code{get_app_dir()}: creates and returns app dir path

\item \code{get_app_colors()}: Set Plot colors

\item \code{create_session_dir()}: Run and Publishing Functions

\item \code{clean_session_dir()}: Run and Publishing Functions

\item \code{save_plot_data()}: keeps plot data generated during session

\item \code{create_dl_link()}: create a dowload link to be displayed to user

\item \code{generate_report()}: Read in template, add the header and write report content to output dir

\item \code{save_plot_image()}: Saving plots for report

\item \code{build_scenario_data()}: This will read from the temp outputs and construct the scenario dataset for the report

\item \code{save_ml_context()}: save ml context (internal)

\item \code{save_ml_results()}: save ml results (internal)

\item \code{exec_ml_restock()}: Run rec model, save results, and return

\item \code{default_ml_params()}: get defaults and scale

\item \code{scale_ml_params()}: scale for shiny sliders

\item \code{unscale_ml_params()}: scale for shiny sliders

}}
