% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml-database.R
\name{ml-database}
\alias{ml-database}
\alias{interp}
\alias{dbGetTransacts90}
\alias{dbGetProductHist}
\alias{dbGetCatPricePts}
\title{Database Interface}
\usage{
interp(txt, cn, oid, sid, sku = NULL)

dbGetTransacts90(oid, sid, sku = NULL, cn = NULL)

dbGetProductHist(oid, sid, sku = NULL, cn = NULL)

dbGetCatPricePts(oid, sid, cn = NULL)
}
\arguments{
\item{txt}{plain text containing query to interpolate}

\item{cn}{a db connection}

\item{oid}{org uuid}

\item{sid}{store uuid}

\item{sku}{one or more skus to generate recommendations for}
}
\description{
These functions are the only functions called by the restock recommendations endpoint that query an external datasource.
}
\section{Functions}{
\itemize{
\item \code{interp()}: interpolate query

\item \code{dbGetTransacts90()}: get recent population data

\item \code{dbGetProductHist()}: get First and Last Sales dates by Sku

\item \code{dbGetCatPricePts()}: get Sales statistics for transactions in the past 90 days by category

}}
