% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltuning-app.R
\name{mltuning-app}
\alias{mltuning-app}
\alias{cleanOutDir}
\alias{createOutDir}
\alias{getDefaultParams}
\alias{saveParams}
\alias{loadParams}
\alias{saveContext}
\alias{savePlotDatasets}
\alias{plotDiag0}
\alias{plotDiag1}
\alias{plotDiag2}
\alias{plotDiag3}
\alias{plotDiag4}
\alias{saveGGPlots}
\alias{buildScenario}
\alias{generateReport}
\alias{getMyColors}
\alias{getMlPlotPanel}
\alias{getBSTheme}
\alias{getSelectionPanel}
\alias{getSlidersRow}
\alias{getMlToggle}
\alias{getMlTuningIndex}
\alias{execMlRestock}
\title{Wrapper Functions for App}
\usage{
cleanOutDir()

createOutDir()

getDefaultParams()

saveParams(...)

loadParams(...)

saveContext(...)

savePlotDatasets(...)

plotDiag0(...)

plotDiag1(...)

plotDiag2(...)

plotDiag3(...)

plotDiag4(...)

saveGGPlots(...)

buildScenario()

generateReport(...)

getMyColors()

getMlPlotPanel()

getBSTheme()

getSelectionPanel()

getSlidersRow()

getMlToggle()

getMlTuningIndex()

execMlRestock(...)
}
\arguments{
\item{...}{arguments passed to internal functions}
}
\description{
Wrapper functions used by app that expose internal package functions
}
\section{Functions}{
\itemize{
\item \code{cleanOutDir()}: app function

\item \code{createOutDir()}: app function

\item \code{getDefaultParams()}: app function

\item \code{saveParams()}: app function

\item \code{loadParams()}: app function

\item \code{saveContext()}: app function

\item \code{savePlotDatasets()}: app function

\item \code{plotDiag0()}: app function

\item \code{plotDiag1()}: app function

\item \code{plotDiag2()}: app function

\item \code{plotDiag3()}: app function

\item \code{plotDiag4()}: app function

\item \code{saveGGPlots()}: app function

\item \code{buildScenario()}: app function

\item \code{generateReport()}: app function

\item \code{getMyColors()}: app ui function

\item \code{getMlPlotPanel()}: app ui function

\item \code{getBSTheme()}: app ui function

\item \code{getSelectionPanel()}: app ui function

\item \code{getSlidersRow()}: app ui function

\item \code{getMlToggle()}: app ui function

\item \code{getMlTuningIndex()}: app ui function

\item \code{execMlRestock()}: app ui function

}}
