% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltuning-plots.R
\name{mltuning-plots}
\alias{mltuning-plots}
\alias{.plot_theme}
\alias{.plot_title_style}
\alias{.parse_fails}
\alias{plot_diagnostic_0}
\alias{plot_diagnostic_1}
\alias{plot_diagnostic_2}
\alias{plot_diagnostic_3}
\alias{plot_diagnostic_4}
\alias{ds_sku_recs_pdata}
\title{Plot functions for MLTuning App}
\usage{
.plot_theme(.colors)

.plot_title_style(ptitle, psubtitle, .colors)

.parse_fails(rec)

plot_diagnostic_0(pdata0, .colors)

plot_diagnostic_1(pdata1, .colors)

plot_diagnostic_2(pdata2, .colors)

plot_diagnostic_3(pdata3, .colors)

plot_diagnostic_4(pdata4, .colors)

ds_sku_recs_pdata(
  oid,
  sid,
  sku,
  ml_trend_conf = 0.85,
  ml_trend_pval = 0.05,
  ml_stock_conf = 0.85,
  ml_stock_pval = 0.05,
  ml_pair_ttest = FALSE,
  ml_pooled_var = TRUE,
  ml_ltmi = 182,
  ml_npom = 14,
  ml_prim = 0.45,
  ml_secd = 0.2,
  ml_ppql = 0.2,
  ml_ppqh = 0.8
)
}
\arguments{
\item{.colors}{colors returned by .hca_colors()}

\item{ptitle}{plot title}

\item{psubtitle}{plot subtitle}

\item{rec}{internal}

\item{pdata0}{plot dataset}

\item{pdata1}{plot dataset}

\item{pdata2}{plot dataset}

\item{pdata3}{plot dataset}

\item{pdata4}{plot dataset}

\item{oid}{see rdscore::restock_rec_ep}

\item{sid}{see rdscore::restock_rec_ep}

\item{sku}{see rdscore::restock_rec_ep}

\item{ml_trend_conf}{see rdscore::restock_rec_ep}

\item{ml_trend_pval}{see rdscore::restock_rec_ep}

\item{ml_stock_conf}{see rdscore::restock_rec_ep}

\item{ml_stock_pval}{see rdscore::restock_rec_ep}

\item{ml_pair_ttest}{see rdscore::restock_rec_ep}

\item{ml_pooled_var}{see rdscore::restock_rec_ep}

\item{ml_ltmi}{see rdscore::restock_rec_ep}

\item{ml_npom}{see rdscore::restock_rec_ep}

\item{ml_prim}{see rdscore::restock_rec_ep}

\item{ml_secd}{see rdscore::restock_rec_ep}

\item{ml_ppql}{see rdscore::restock_rec_ep}

\item{ml_ppqh}{see rdscore::restock_rec_ep}
}
\description{
Plot functions for MLTuning App
}
\section{Functions}{
\itemize{
\item \code{.plot_theme()}: provides the base theme for plots

\item \code{.plot_title_style()}: Style the plot title/subtitle

\item \code{.parse_fails()}: Parse failed skus to delineate between failures and uncertain recs

\item \code{plot_diagnostic_0()}: diagnostic plot

\item \code{plot_diagnostic_1()}: diagnostic plot

\item \code{plot_diagnostic_2()}: diagnostic plot

\item \code{plot_diagnostic_3()}: diagnostic plot

\item \code{plot_diagnostic_4()}: diagnostic plot

\item \code{ds_sku_recs_pdata()}: Run recommendations then Build plot datasets and generate plots

}}
