% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltuning-ui.R
\name{mltuning-ui}
\alias{mltuning-ui}
\alias{uiBootstrapTheme}
\alias{uiSelectionFilts}
\alias{uiSelectionStats}
\alias{uiModelPlotPanel}
\alias{.slider_model_sales}
\alias{.slider_model_stock}
\alias{.slider_price_qrtls}
\alias{.slider_product_cls}
\alias{.slider_menu_period}
\alias{.toggle_model_ttest}
\title{Ui elements of MLTuning App}
\usage{
uiBootstrapTheme()

uiSelectionFilts(width = 12, inline = TRUE)

uiSelectionStats(width = 12, .colors)

uiModelPlotPanel(width = 12)

.slider_model_sales(width = "100\%", inline = FALSE)

.slider_model_stock(width = "100\%", inline = FALSE)

.slider_price_qrtls(width = "100\%", inline = FALSE)

.slider_product_cls(width = "100\%", inline = FALSE)

.slider_menu_period(width = "100\%", inline = FALSE)

.toggle_model_ttest(width = "150px", inline = TRUE)
}
\arguments{
\item{width}{width}

\item{inline}{default TRUE}

\item{.colors}{hca colors}
}
\description{
Ui elements of MLTuning App
}
\section{Functions}{
\itemize{
\item \code{uiBootstrapTheme()}: bootstrap theme for app

\item \code{uiSelectionFilts()}: statistics panel

\item \code{uiSelectionStats()}: statistics panel

\item \code{uiModelPlotPanel()}: model diagnostic plot panel

\item \code{.slider_model_sales()}: slider

\item \code{.slider_model_stock()}: slider

\item \code{.slider_price_qrtls()}: slider

\item \code{.slider_product_cls()}: slider

\item \code{.slider_menu_period()}: slider

\item \code{.toggle_model_ttest()}: ml toggle

}}
