% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltuning-utils.R
\name{mltuning-utils}
\alias{mltuning-utils}
\alias{get_index}
\alias{.hca_colors}
\alias{createOutputDir}
\alias{cleanOutputDir}
\alias{savePlotData}
\alias{generate_report}
\alias{save_ggplots}
\alias{build_scenario_data}
\alias{save_ml_context}
\alias{exec_ml_restock}
\alias{default_params}
\alias{scale_params}
\alias{unscale_params}
\alias{load_params}
\alias{save_params}
\title{MLTuning App Utils}
\usage{
get_index()

.hca_colors()

createOutputDir()

cleanOutputDir()

savePlotData(results)

generate_report(org, store)

save_ggplots(p0, p1, p2, p3, p4)

build_scenario_data()

save_ml_context(oid, sid, sku_data, ml_args)

exec_ml_restock(oid, sid, skus, ml_args)

default_params()

scale_params(ll)

unscale_params(ll)

load_params(oid, sid)

save_params(oid, sid, args)
}
\arguments{
\item{results}{results from shiny app}

\item{org}{org name}

\item{store}{store name}

\item{p0}{plot}

\item{p1}{plot}

\item{p2}{plot}

\item{p3}{plot}

\item{p4}{plot}

\item{oid}{org id}

\item{sid}{store id}

\item{sku_data}{sku table}

\item{ml_args}{model args}

\item{skus}{skus to get recs for}

\item{ll}{argument list to scale or unscale values}

\item{args}{args to send to the db}
}
\description{
MLTuning App Utils
}
\section{Functions}{
\itemize{
\item \code{get_index()}: get primary app data

\item \code{.hca_colors()}: Set Plot colors

\item \code{createOutputDir()}: Run and Publishing Functions

\item \code{cleanOutputDir()}: Run and Publishing Functions

\item \code{savePlotData()}: keeps plot data generated during session

\item \code{generate_report()}: Read in template, add the header and write report content to output dir

\item \code{save_ggplots()}: Saving plots for report

\item \code{build_scenario_data()}: This will read from the temp outputs and construct the scenario dataset for the report

\item \code{save_ml_context()}: save ml context (internal)

\item \code{exec_ml_restock()}: Run rec model, save results, and return

\item \code{default_params()}: get defaults and scale

\item \code{scale_params()}: scale for shiny sliders

\item \code{unscale_params()}: scale for shiny sliders

\item \code{load_params()}: get default model params by location from the db

\item \code{save_params()}: save custom model params by location to the db

}}
