% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestockML-package.R
\docType{package}
\name{RestockML-package}
\alias{RestockML}
\alias{RestockML-package}
\title{RestockML: Inventory Restock Recommendations Model}
\description{
Package containing inventory restock recommendation model and accompanied app for exploring and fine-tuning the model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-data-science/RestockML}
  \item Report bugs at \url{https://github.com/r-data-science/RestockML/issues}
}

}
\author{
\strong{Maintainer}: Bobby Fatemi \email{bfatemi07@gmail.com}

}
\keyword{internal}
