% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{test-helpers}
\alias{test-helpers}
\alias{get_db_params}
\alias{clear_db_params}
\alias{compare_report}
\title{Internal Testing Helper Functions}
\usage{
get_db_params(oid, sid)

clear_db_params(oid, sid)

compare_report(old, new)
}
\arguments{
\item{oid}{org uuid}

\item{sid}{store uuid}

\item{old}{filepath to old document}

\item{new}{filepath to new document}
}
\description{
Unexported functions used in package unit tests
}
\section{Functions}{
\itemize{
\item \code{get_db_params()}: Function to retrieve database saved parameters for checking in unit tests

\item \code{clear_db_params()}: Function to clear any database parameters saved during unit tests

\item \code{compare_report()}: Function to compare downloaded report in unit testing

}}
