% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{app-utils}
\alias{app-utils}
\alias{is_ci}
\alias{is_testing}
\alias{get_app_dir}
\alias{get_app_colors}
\alias{create_session_dir}
\alias{clear_session_dir}
\alias{generate_report}
\alias{build_plot_data}
\alias{save_plot_data}
\alias{build_plot_objects}
\alias{save_plot_objects}
\alias{build_ml_scenario}
\alias{save_ml_scenario}
\alias{build_ml_context}
\alias{save_ml_context}
\alias{build_ml_recs}
\alias{save_ml_recs}
\alias{default_ml_params}
\alias{scale_ml_params}
\alias{unscale_ml_params}
\title{App Utils}
\usage{
is_ci()

is_testing()

get_app_dir()

get_app_colors()

create_session_dir()

clear_session_dir()

generate_report(file)

build_plot_data(recs)

save_plot_data(results)

build_plot_objects(results)

save_plot_objects(plots)

build_ml_scenario(results, context)

save_ml_scenario(scenario)

build_ml_context(oid, sid, index, ml_args)

save_ml_context(context)

build_ml_recs(oid, sid, skus, ml_args)

save_ml_recs(recs)

default_ml_params()

scale_ml_params(ll)

unscale_ml_params(ll)
}
\arguments{
\item{file}{file name for report download passed from app}

\item{recs}{rec results from shiny app}

\item{results}{results from shiny app}

\item{plots}{list of diagnostic plots}

\item{context}{model context}

\item{scenario}{report scenario}

\item{oid}{org id}

\item{sid}{store id}

\item{index}{filtered products table from app}

\item{ml_args}{model args}

\item{skus}{skus to get recs for}

\item{ll}{argument list to scale or unscale values}
}
\description{
Functions required to execute and facililate an application user session
}
\section{Functions}{
\itemize{
\item \code{is_ci()}: returns TRUE if called on CI

\item \code{is_testing()}: returns TRUE if called while testing

\item \code{get_app_dir()}: creates and returns app dir path

\item \code{get_app_colors()}: Set Plot colors

\item \code{create_session_dir()}: Run and Publishing Functions

\item \code{clear_session_dir()}: Run and Publishing Functions

\item \code{generate_report()}: Generate report and return download link

\item \code{build_plot_data()}: build plot data with result of rdscore::restock_rec_ep

\item \code{save_plot_data()}: keeps plot data generated during session

\item \code{build_plot_objects()}: get list of diagnostic plots given processed recs results

\item \code{save_plot_objects()}: Saving plots for report

\item \code{build_ml_scenario()}: build model scenario

\item \code{save_ml_scenario()}: save model scenario

\item \code{build_ml_context()}: build ml context (internal)

\item \code{save_ml_context()}: save ml context (internal)

\item \code{build_ml_recs()}: Run recs model, save results, and return

\item \code{save_ml_recs()}: save ml results (internal)

\item \code{default_ml_params()}: get defaults and scale

\item \code{scale_ml_params()}: scale for shiny sliders

\item \code{unscale_ml_params()}: scale for shiny sliders

}}
