% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsapps-package.R
\docType{package}
\name{rdsapps-package}
\alias{rdsapps}
\alias{rdsapps-package}
\title{rdsapps: Inventory Restock Recommendations Model}
\description{
Package containing inventory restock recommendation model and accompanied app for exploring and fine-tuning the model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-data-science/rdsapps}
  \item Report bugs at \url{https://github.com/r-data-science/rdsapps/issues}
}

}
\author{
\strong{Maintainer}: Bobby Fatemi \email{bfatemi07@gmail.com}

}
\keyword{internal}
