% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch-app.R
\name{run-app}
\alias{run-app}
\alias{runAppRecModel}
\title{Launch Package Apps}
\usage{
runAppRecModel(SAVE_PLOTDATA = FALSE, CLEAN_OUTPUTS = TRUE)
}
\arguments{
\item{SAVE_PLOTDATA}{Default is FALSE. Whether to keep plot input datasets generated during app session. Useful for plot development.}

\item{CLEAN_OUTPUTS}{Default is False. Whether to delete the session output directory. Useful for debuging}
}
\description{
Launch Package Apps
}
\section{Functions}{
\itemize{
\item \code{runAppRecModel()}: run MLTuning app

}}
